% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributable_risk.R
\name{attributable_risk}
\alias{attributable_risk}
\title{attributable_risk function}
\usage{
attributable_risk(
  x,
  basis,
  cases,
  model,
  dir = "back",
  average = NULL,
  tot = TRUE,
  cen,
  range = NULL,
  n_sample = 1000,
  seed = 0L
)
}
\arguments{
\item{seed}{}
}
\description{
attributable_risk function
}
