% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcrosspred.R
\name{bcrosspred}
\alias{bcrosspred}
\title{bcrosspred
This function generates predictions from bayesian distributed lag linear (DLMs) and distributed lag non-linear models (DLNM).}
\usage{
bcrosspred(
  basis,
  model = NULL,
  coef = NULL,
  at = NULL,
  from = NULL,
  to = NULL,
  by = NULL,
  lag,
  bylag = 1,
  cen = NULL,
  cumul = FALSE,
  n_sample = 100
)
}
\arguments{
\item{n_sample}{}
}
\description{
bcrosspred
This function generates predictions from bayesian distributed lag linear (DLMs) and distributed lag non-linear models (DLNM).
}
