% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimum_risk.R
\name{minimum_risk}
\alias{minimum_risk}
\title{minimum_risk: function to obtain the minimum exposure value in which there is the minimum overall risk effect.}
\usage{
minimum_risk(
  basis,
  model = NULL,
  at = NULL,
  from = NULL,
  to = NULL,
  by = NULL,
  n_sample = 1000,
  seed = 0L
)
}
\arguments{
\item{seed}{}
}
\description{
minimum_risk: function to obtain the minimum exposure value in which there is the minimum overall risk effect.
}
