% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_exposure.R
\name{optimal_exposure}
\alias{optimal_exposure}
\title{Optimal effect exposure value estimation}
\usage{
optimal_exposure(
  object,
  basis,
  at = NULL,
  from = NULL,
  to = NULL,
  by = NULL,
  which = "min"
)
}
\arguments{
\item{object}{A fitted \code{"bdlnm"} class object returned by \link{bdlnm}.}

\item{basis}{A DLNM basis object produced by \code{dlnm}. It can be one of \link[dlnm:crossbasis]{dlnm::crossbasis} or \link[dlnm:onebasis]{dlnm::onebasis}.}

\item{at}{Values (or matrix) of the predictor at which to predict; can be \code{NULL} and reconstructed from \code{from}, \code{to}, \code{by} and the basis attributes.}

\item{from, to, by}{Optional numeric used to construct \code{at} when not provided.}

\item{which}{Character string specifying if the optimal exposure value is chosen as the one that minimizes the effect or as the one that maximizes it. By default is \code{"min"}.}
}
\value{
A list of class \code{optimal_exposure} containing:
\itemize{
\item \code{est}: numeric vector with the optimal exposure value for each posterior sample
\item \code{summary}: data frame with summary statistics for these optimal values
}

The returned object has attribute \code{xvar} containing the exposure grid used.
}
\description{
The function finds the exposure value that optimizes the overall effect for each posterior sample, together with a small summary of these values. The user can decide to choose the optimal exposure as the value that has either the minimum or the maximum estimated effect. In case of finding the minimum exposure effect in the context of the relationship between temperature and mortality, this value is called Minimum Mortality Temperature (MMT).
}
\examples{
# Set exposure-response and lag-response spline parameters
 dlnm_var <- list(
   var_prc = c(10, 75, 90),
   var_fun = "ns",
   lag_fun = "ns",
   max_lag = 21,
   lagnk = 3
 )


# Set cross-basis parameters
 argvar <- list(fun = dlnm_var$var_fun,
                knots = stats::quantile(london$tmean,
                                 dlnm_var$var_prc/100, na.rm = TRUE),
                Bound = range(london$tmean, na.rm = TRUE))

 arglag <- list(fun = dlnm_var$lag_fun,
                knots = dlnm::logknots(dlnm_var$max_lag, nk = dlnm_var$lagnk))

 # Create crossbasis
 cb <- dlnm::crossbasis(london$tmean, lag = dlnm_var$max_lag, argvar, arglag)

 # Seasonality of mortality time series
 seas <- splines::ns(london$date, df = round(8 * length(london$date) / 365.25))

 # Prediction values (equidistant points)
 temp <- round(seq(min(london$tmean), max(london$tmean), by = 0.1), 1)

 # Fit the model
 mod <- bdlnm(mort_75plus ~ cb + factor(dow) + seas, basis = cb, data = london, family = "poisson")

 # Find minimum risk exposure value
 mmt <- optimal_exposure(mod, cb, at = temp)

}
