% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.bcrosspred.R
\name{plot.bcrosspred}
\alias{plot.bcrosspred}
\title{plot.bcrosspred
This function generates plots for bayesian distributed lag non-linear model's predictions stored in a bcrosspred object.}
\usage{
\method{plot}{bcrosspred}(
  x,
  ptype,
  var = NULL,
  lag = NULL,
  ci = "area",
  ci.arg,
  ci.level = x$ci.level,
  cumul = FALSE,
  exp = NULL,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
plot.bcrosspred
This function generates plots for bayesian distributed lag non-linear model's predictions stored in a bcrosspred object.
}
