% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.min.risk.R
\name{plot.min.risk}
\alias{plot.min.risk}
\title{Plot posterior distribution of minimum-risk exposure values}
\usage{
\method{plot}{min.risk}(x, line.arg = NULL, ...)
}
\arguments{
\item{x}{Object of class \code{min.risk} as returned by \link{minimum_risk}.}

\item{line.arg}{List of graphical arguments for the plotting of the median vertical line passed to \link[graphics:abline]{graphics::abline}.}

\item{...}{Additional graphical parameters passed to \link{hist}.}
}
\description{
Plot the posterior distribution of the minimum-risk exposure values returned by \code{minimum_risk()}.
}
\details{
Plot an histogram showing the posterior distribution of the minimum-risk exposure values returned by \code{minimum_risk()}. The original prediction grid will be used as axis breaks. A dashed red vertical line marks the posterior median (0.5 quantile when available).
}
\examples{

# Set exposure-response and lag-response spline parameters
 dlnm_var <- list(
   var_prc = c(10, 75, 90),
   var_fun = "ns",
   lag_fun = "ns",
   max_lag = 21,
   lagnk = 3
 )


# Set cross-basis parameters
 argvar <- list(fun = dlnm_var$var_fun,
                knots = stats::quantile(london$tmean,
                                 dlnm_var$var_prc/100, na.rm = TRUE),
                Bound = range(london$tmean, na.rm = TRUE))

 arglag <- list(fun = dlnm_var$lag_fun,
                knots = dlnm::logknots(dlnm_var$max_lag, nk = dlnm_var$lagnk))

 # Create crossbasis
 cb <- dlnm::crossbasis(london$tmean, lag = dlnm_var$max_lag, argvar, arglag)

 # Seasonality of mortality time series
 seas <- splines::ns(london$date, df = round(8 * length(london$date) / 365.25))

 # Prediction values (equidistant points)
 temp <- round(seq(min(london$tmean), max(london$tmean), by = 0.1), 1)

 # Fit the model
 mod <- bdlnm(mort_75plus ~ cb + factor(dow) + seas, basis = cb, data = london,
 family = "poisson")

 # Find minimum risk exposure value
 mmt <- minimum_risk(mod, cb, at = temp)

 # Plot
 plot(mmt, xlab = "Temperature (ºC)",
 main = paste0("MMT (Median = ", round(mmt$min.summary[["0.5quant"]], 1), "ºC)"))


}
