% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.optimal_exposure.R
\name{plot.optimal_exposure}
\alias{plot.optimal_exposure}
\title{Plot posterior distribution of optimal effect exposure values}
\usage{
\method{plot}{optimal_exposure}(x, show_median = TRUE, vline.arg = NULL, ...)
}
\arguments{
\item{x}{An object of class \code{"optimal_exposure"} returned by \code{\link[=optimal_exposure]{optimal_exposure()}}.}

\item{show_median}{Logical. If \code{TRUE} (default) the function draws a vertical line for the median of all the posterior samples. If \code{FALSE} it doesn't draw any additional line.}

\item{vline.arg}{Optional list of graphical arguments passed to \code{\link[graphics:abline]{graphics::abline()}} when drawing the median vertical line.}

\item{...}{Optional graphical parameters passed to \code{\link[graphics:hist]{graphics::hist()}}.}
}
\description{
It plots a histogram of the posterior distribution of the optimal effect exposure values returned by \code{\link[=optimal_exposure]{optimal_exposure()}}.
}
\details{
The histogram uses the original prediction grid in \code{attr(object, "xvar")} as the x-axis values, ensuring that the bars align with prediction exposure values. The function plots the posterior distribution of the optimal exposure values (stored in \code{x$est}) and highlights the posterior median across samples (stored in \code{x$summary[["0.5quant"]]}) with a vertical line if \code{show_median = TRUE}. Use \code{vline.arg} to change the appearance of that line passed to \code{\link[graphics:abline]{graphics::abline()}} and \code{...} to change the graphical parameters of the histogram passed to \code{\link[graphics:hist]{graphics::hist()}} (to control axis labels, title, colours, etc.). See the original functions for a complete list of the arguments. Some arguments, if not specified, are set to different default values than the original functions.
}
\examples{

# Set exposure-response and lag-response spline parameters
 dlnm_var <- list(
   var_prc = c(10, 75, 90),
   var_fun = "ns",
   lag_fun = "ns",
   max_lag = 21,
   lagnk = 3
 )


# Set cross-basis parameters
 argvar <- list(fun = dlnm_var$var_fun,
                knots = stats::quantile(london$tmean,
                                 dlnm_var$var_prc/100, na.rm = TRUE),
                Bound = range(london$tmean, na.rm = TRUE))

 arglag <- list(fun = dlnm_var$lag_fun,
                knots = dlnm::logknots(dlnm_var$max_lag, nk = dlnm_var$lagnk))

 # Create crossbasis
 cb <- dlnm::crossbasis(london$tmean, lag = dlnm_var$max_lag, argvar, arglag)

 # Seasonality of mortality time series
 seas <- splines::ns(london$date, df = round(8 * length(london$date) / 365.25))

 # Prediction values (equidistant points)
 temp <- round(seq(min(london$tmean), max(london$tmean), by = 0.1), 1)

 # Fit the model
 mod <- bdlnm(mort_75plus ~ cb + factor(dow) + seas, basis = cb, data = london,
 family = "poisson")

 # Find minimum risk exposure value
 mmt <- optimal_exposure(mod, cb, at = temp)

 # Plot
 plot(mmt, xlab = "Temperature (ºC)",
 main = paste0("MMT (Median = ", round(mmt$summary[["0.5quant"]], 1), "ºC)"))


}
\references{
Quijal-Zamorano M, Martinez-Beneito MA, Ballester J, Marí-Dell’Olmo M. Spatial Bayesian distributed lag non-linear models (SB-DLNM) for small-area exposure-lag-response epidemiological modelling. International Journal of Epidemiology. 2024;53(3):dyae061.

Gasparrini A (2011). Distributed lag linear and non-linear models in R: the package dlnm. Journal of Statistical Software, 43(8), 1–20.

Armstrong B. Models for the relationship between ambient temperature and daily mortality. Epidemiology. 2006;17(6):624-31.
}
\seealso{
\code{\link[=optimal_exposure]{optimal_exposure()}} to estimate exposure values that optimize the predicted effect for a \code{"bdlnm"} object.
}
\author{
Pau Satorra, Marcos Quijal-Zamorano.
}
