% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary_bdlnm}
\alias{summary_bdlnm}
\alias{summary.bdlnm}
\alias{summary.bcrosspred}
\alias{summary.optimal_exposure}
\title{summary methods for bdlnm objects}
\usage{
\method{summary}{bdlnm}(object, ...)

\method{summary}{bcrosspred}(object, ...)

\method{summary}{optimal_exposure}(object, ...)
}
\arguments{
\item{object}{An object created using bdlnm functions}

\item{...}{Not used}
}
\description{
summary methods for bdlnm objects
}
\keyword{internal}
