% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributable.R
\name{attributable}
\alias{attributable}
\title{Calculate attributable number and fractions from Bayesian distributed-lag models (B-DLNM).}
\usage{
attributable(
  object,
  basis,
  data,
  name_date = NULL,
  name_exposure,
  name_cases = NULL,
  name_filter = NULL,
  dir = "back",
  cen,
  range = NULL,
  average = TRUE
)
}
\arguments{
\item{object}{A fitted \code{"bdlnm"} class object returned by \link{bdlnm}.}

\item{basis}{A DLNM crossbasis object produced by \code{dlnm}. It has to be of class \link[dlnm:crossbasis]{dlnm::crossbasis}.}

\item{data}{A dataframe containing the time series of exposure values (see \code{name_exposure}) and number of cases (see \code{name_cases}). Ensure that the time series are provided on a regular basis for attributable measures to be calculated properly. It can also include a column with the measured time point (see \code{name_date}) or a column with a \code{0/1} indicator that filters the calculation of attributable measures for specific time periods (see \code{name_filter}).}

\item{name_date}{A character with the name of the column with the date (class \code{Date} or \code{POSICXt}) of the time series measurement (optional).}

\item{name_exposure}{A character with the name of the column with the time series exposure values.}

\item{name_cases}{A character with the name of the column with the time series case values. If not provided, only attributable fractions per time point can be calculated.}

\item{name_filter}{A character with the name of the column with the indicator that can filter the time points in which to calculate attributable measures (optional).}

\item{dir}{Character; "back" (default) or "forw" direction in the lag dimension for calculating attributable numbers and fractions.}

\item{cen}{Numeric scalar; centering value for predictions. If missing the function will attempt to read it from attr(basis, "argvar")$cen.}

\item{range}{Optional numeric vector of range 2. It gives the exposure value range for which attributable numbers and fractions will be calculated.}

\item{average}{Logical (default is TRUE). When TRUE the function uses average (lag-averaged) contributions to calculate attributable risks. Only applied in the forward perspective.}
}
\value{
A list with elements:
\itemize{
\item \code{af}: attributable fraction (AF) per time point
\item \code{an}: attributable number (AN) per time point
\item \code{aftotal}: total attributable fraction (AF) across all period
\item \code{antotal}: total attributable number (AN) across all period
\item \code{af.summary}: data frame with summary statistics for AF per time point
\item \code{an.summary}: data frame with summary statistics for AN per time point
\item \code{aftotal.summary}: data frame with summary statistics for total AF
\item \code{antotal.summary}: data frame with summary statistics for total AN
}
}
\description{
It computes the attributable number (AN) and fraction (AF) from a fitted B-DLNM returned by \code{bdlnm()}.
}
\examples{

# Filter the dataset to reduce computational time:

# Exposure-response and lag-response spline parameters
dlnm_var <- list(
  var_prc = c(10, 75, 90),
  var_fun = "ns",
  lag_fun = "ns",
  max_lag = 21,
  lagnk = 3
)

# Cross-basis parameters
argvar <- list(fun = dlnm_var$var_fun,
               knots = stats::quantile(london$tmean,
                                dlnm_var$var_prc/100, na.rm = TRUE),
               Bound = range(london$tmean, na.rm = TRUE))

arglag <- list(fun = dlnm_var$lag_fun,
               knots = dlnm::logknots(dlnm_var$max_lag, nk = dlnm_var$lagnk))

# Create crossbasis
cb <- dlnm::crossbasis(london$tmean, lag = dlnm_var$max_lag, argvar, arglag)

# Seasonality of mortality time series
seas <- splines::ns(london$date, df = round(8 * length(london$date) / 365.25))

# Prediction values (equidistant points)
temp <- round(seq(min(london$tmean), max(london$tmean), by = 0.1), 1)

# Model

mod <- bdlnm(mort_75plus ~ cb + factor(dow) + seas,
            basis = cb,
             data = london,
             family = "poisson",
             sample.arg = list(seed = 432))

# Predict
cpred <- bcrosspred(mod, cb, at = temp)

# compute centering (MMT) using optimal_exposure
mmt <- optimal_exposure(mod, cb, at = temp)
cen <- mmt$summary[["0.5quant"]]

# Attributable numbers and fractions (using the backwards algorithm):
ar <- attributable(mod, cb, london, name_date = "date",
name_exposure = "tmean", name_cases = "mort_75plus", cen = cen, dir = "back")



}
