% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdlnm.R
\name{bdlnm}
\alias{bdlnm}
\title{Fit a bayesian distributed lag non-linear model (B-DLNM)}
\usage{
bdlnm(
  formula,
  basis,
  data,
  family = "gaussian",
  sample.arg = list(n = 1000, seed = 0L),
  ci.level = 0.95,
  ...
)
}
\arguments{
\item{formula}{A model formula (as for \link[INLA:inla]{INLA::inla}).}

\item{basis}{A DLNM basis object produced by \code{dlnm}. It can be one of \link[dlnm:crossbasis]{dlnm::crossbasis} or \link[dlnm:onebasis]{dlnm::onebasis}.}

\item{data}{A data frame used for the model fitted with \link[INLA:inla]{INLA::inla}.}

\item{family}{Family name passed to \link[INLA:inla]{INLA::inla} (default \code{"gaussian"}).}

\item{sample.arg}{List containing the arguments passed to \link[INLA:posterior.sample]{INLA::inla.posterior.sample}. Default sets \code{n = 1000} as the number of samples and \code{seed = 0L} (set the seed at 'random').}

\item{ci.level}{Numeric in \verb{(0,1)} giving the credible interval level (default \code{0.95}).}

\item{...}{Additional arguments passed to \link[INLA:inla]{INLA::inla}.}
}
\value{
A list with components:
\itemize{
\item \code{model}: The fitted \code{INLA} model.
\item \code{coef}: A matrix with posterior samples for the basis coefficients
}
}
\description{
Fit a distributed lag non-linear model (DLNM) that includes a basis object from \code{dlnm}. It uses \code{INLA} to fit the bayesian model and to extract posterior samples of the basis coefficients.
}
\details{
If the basis indicates some lag, the first \code{max(lags)} rows of \code{data} are set to \code{NA} to omit the basis first missing rows for model fitting.
}
\examples{

# Set exposure-response and lag-response spline parameters
 dlnm_var <- list(
   var_prc = c(10, 75, 90),
   var_fun = "ns",
   lag_fun = "ns",
   max_lag = 21,
   lagnk = 3
 )


# Set cross-basis parameters
 argvar <- list(fun = dlnm_var$var_fun,
                knots = stats::quantile(london$tmean,
                                 dlnm_var$var_prc/100, na.rm = TRUE),
                Bound = range(london$tmean, na.rm = TRUE))

 arglag <- list(fun = dlnm_var$lag_fun,
                knots = dlnm::logknots(dlnm_var$max_lag, nk = dlnm_var$lagnk))

 # Create crossbasis
 cb <- dlnm::crossbasis(london$tmean, lag = dlnm_var$max_lag, argvar, arglag)

 # Seasonality of mortality time series
 seas <- splines::ns(london$date, df = round(8 * length(london$date) / 365.25))

 # Prediction values (equidistant points)
 temp <- round(seq(min(london$tmean), max(london$tmean), by = 0.1), 1)

 # Fit the model
 mod <- bdlnm(mort_75plus ~ cb + factor(dow) + seas, basis = cb, data = london, family = "poisson")


}
