% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.bcrosspred.R
\name{plot.bcrosspred}
\alias{plot.bcrosspred}
\title{Plot predicted effects from bayesian DLNM (B-DLNM)}
\usage{
\method{plot}{bcrosspred}(
  x,
  ptype,
  var = NULL,
  lag = NULL,
  ci = "area",
  ci.arg,
  ci.level = x$ci.level,
  cumul = FALSE,
  exp = NULL,
  ...
)
}
\arguments{
\item{x}{An object of class \code{"bcrosspred"} (returned by \link{bcrosspred}).}

\item{ptype}{Character. Plot type: one of \code{"slices"}, \code{"3d"}, \code{"contour"} or \code{"overall"}.}

\item{var}{Numeric vector of predictor values for slices plot type (maximum length of 4).}

\item{lag}{Numeric vector of lags for slices plot type (maximum length of 4).}

\item{ci}{Character. How to plot credible intervals: one of \code{"area"}, \code{"bars"}, \code{"lines"}, \code{"sampling"} (draw a line for each posterior sample) or \code{"none"}. Default: \code{"area"}.}

\item{ci.arg}{List of graphical arguments for the plotting of credible intervals passed to \link[graphics:lines]{graphics::lines}.}

\item{ci.level}{Numeric in (0,1). Credible interval level. Default will take the \code{ci.level} specified in \code{x}.}

\item{cumul}{Logical. If \code{TRUE} it will plot cumulative predictions (requires cumulative predictions in \code{x}).}

\item{exp}{Logical. When \code{TRUE} plot on the exponentiated effects. Default will be \code{TRUE} only if \code{model.link} is \code{log/logit} in \code{x}.}

\item{...}{Additional graphical arguments passed to base plotting functions.}
}
\description{
Plot lag-specific, overall, slices, contour or 3D representations for predicted effects generated by \code{bcrosspred()}.
}
\examples{
# Set exposure-response and lag-response spline parameters
 dlnm_var <- list(
   var_prc = c(10, 75, 90),
   var_fun = "ns",
   lag_fun = "ns",
   max_lag = 21,
   lagnk = 3
 )


# Set cross-basis parameters
 argvar <- list(fun = dlnm_var$var_fun,
                knots = stats::quantile(london$tmean,
                                 dlnm_var$var_prc/100, na.rm = TRUE),
                Bound = range(london$tmean, na.rm = TRUE))

 arglag <- list(fun = dlnm_var$lag_fun,
                knots = dlnm::logknots(dlnm_var$max_lag, nk = dlnm_var$lagnk))

 # Create crossbasis
 cb <- dlnm::crossbasis(london$tmean, lag = dlnm_var$max_lag, argvar, arglag)

 # Seasonality of mortality time series
 seas <- splines::ns(london$date, df = round(8 * length(london$date) / 365.25))

 # Prediction values (equidistant points)
 temp <- seq(round(min(london$tmean), 1), round(max(london$tmean), 1), by = 0.1)

 # Fit the model
 mod <- bdlnm(mort_75plus ~ cb + factor(dow) + seas, basis = cb, data = london, family = "poisson")

 # Prediction
 cpred <- bcrosspred(x, cb, at = temp)

 # Perform the plots:

 #Overall
  plot(cpred, "overall", xlab = "Temperature (ºC)", ylab = "Relative Risk", col = 4,
  main="Overall", log = "y")
#3-D plot
  plot(cpred, zlab = "Relative risk", col = 4, lphi = 60, cex.axis = 0.6,
  xlab = "Temperature (ºC)", main = "3D graph of temperature effect")
  #Contour
 plot(cpred, "contour", xlab = "Temperature (ºC)", ylab = "Lag", main="Contour plot")
 #Slices (for a high temperature)
 htemp <- 23
 plot(cpred , "slices", var = htemp, col=3, ylab="RR",
 main=paste0("Association for a high temperature (", htemp, "ºC)"))
 #Slices (for lag 0)
 plot(cpred , "slices", lag = 0, col=4, ylab="RR", main=paste0("Association at Lag 0"))


}
