% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.bcrosspred.R
\name{plot.bcrosspred}
\alias{plot.bcrosspred}
\title{Plot the predicted effects from a Bayesian distributed-lag model (B-DLNM).}
\usage{
\method{plot}{bcrosspred}(
  x,
  ptype,
  var = NULL,
  lag = NULL,
  ci = "area",
  ci.arg,
  ci.level = x$ci.level,
  cumul = FALSE,
  exp = NULL,
  ...
)
}
\arguments{
\item{x}{An object of class \code{"bcrosspred"} returned by \code{\link[=bcrosspred]{bcrosspred()}}.}

\item{ptype}{Character. Plot type: one of \code{"slices"}, \code{"overall"}, \code{"3d"} or \code{"contour"}.}

\item{var}{Numeric vector of predictor values (exposure) for the \code{"slices"} plot type (maximum length of 4).}

\item{lag}{Numeric vector of lags for the \code{"slices"} plot type (maximum length of 4).}

\item{ci}{Character. How to plot credible intervals: one of \code{"area"}, \code{"bars"}, \code{"lines"}, \code{"sampling"} (draw a line for each posterior sample) or \code{"none"}. Default: \code{"area"}.}

\item{ci.arg}{List of graphical arguments for the plotting of credible intervals passed to base plotting functions (see \code{Details}).}

\item{ci.level}{Numeric in \verb{(0,1)}. Credible interval level. Default is taken from \code{x$ci.level}.}

\item{cumul}{Logical. If \code{TRUE}, plot incremental cumulative predictions (requires cumulative predictions in \code{x}).}

\item{exp}{Logical. Set to \code{TRUE} to plot exponentiated effects (relative risks) and to \code{FALSE} to not plot them. The default is \code{NULL} in which the function uses \code{x$model.link} to automatically detect wheteher to use relative risks (when link is \code{log} or \code{logit}).}

\item{...}{Additional graphical arguments passed to base plotting functions (see \code{Details}).}
}
\description{
It plots the lag-specific, overall, slices, contour or 3D representations for predicted effects generated by \code{\link[=bcrosspred]{bcrosspred()}}.
}
\details{
The function supports different visualizations for posterior effects predicted by \code{\link[=bcrosspred]{bcrosspred()}} specified in \code{ptype}:
\itemize{
\item \code{"slices"}: (optionally multi-panel) line plot of the predicted effect over lag(s) for selected exposure value(s) or over exposure value(s) for selected lag(s). Use \code{var} and/or \code{lag} to specify which slices to draw (at least one must be supplied). At most 4 slices of each type are allowed to keep layouts readable. See \code{\link[=plot.default]{plot.default()}}, \code{\link[=lines]{lines()}} and \code{\link[=points]{points()}} for information on additional graphical arguments.
\item \code{"overall"}: plot the predicted overall cumulative effect over the whole lag period for each exposure value in the prediction grid. See \code{\link[=plot.default]{plot.default()}}, \code{\link[=lines]{lines()}} and \code{\link[=points]{points()}} for information on additional graphical arguments.
\item \code{"3d"}: 3D plot of the surface of the predicted effect for each exposure and lag. Uses the median of the posterior samples. Not meaningful for single lag models. Additional graphical arguments can be included, such as \code{theta} or \code{phi} (perspective), \code{border} or \code{shade} (surface), \code{xlab}, \code{ylab} or \code{zlab} (axis labelling) and \code{col.} See \code{\link[=persp]{persp()}} for additional information.
\item \code{"contour"}: filled contour of the surface of the predicted effect for each exposure and lag. Uses the median of the posterior samples. Not meaningful for single lag models. Additional graphical arguments can be included, such as \code{plot.title}, \code{plot.axes} or \code{key.title} for titles and axis and key labelling. See \code{\link[=filled.contour]{filled.contour()}} for additional information.
}

The function will call the specified original base plotting functions for each different \code{ptype}. Via the argument \code{...} the user can change the graphical parameters that will be passed to these functions. See the original functions for a complete list of the arguments. Some arguments, if not specified, are set to different default values than the original functions.

Credible intervals will only be drawn for \code{ptype} equal to \code{"slices"} or \code{"overall"}. The type of credible interval will be given by the \code{ci} argument, with options \code{"area"} (default), \code{"bars"}, \code{"lines"}, \code{"sampling"} (draw a line for each posterior sample) or \code{"none"} (no credible intervals). Their appearance may be modified through \code{ci.arg}, a list of arguments passed to to low-level plotting functions: \code{\link[=polygon]{polygon()}} for \code{"area"}, \code{\link[=segments]{segments()}} for \code{"bars"} and \code{\link[=lines]{lines()}} for \code{"lines"}. See the original functions for a complete list of the arguments. As above, some unspecified arguments are set to different default values. The credible interval level will be given by \code{ci.level} or inferred from \code{x$ci.level} by default.

If exponentiated effects (relative risks) are specified (\code{exp = TRUE}) or auto-detected if \code{x$model.link} is \code{log} or \code{logit}, the function will use the predicted relative risks stored in \code{x$matRRfit}, \code{x$allRRfit} or \code{x$cumRRfit} (if \code{cumul=TRUE}).
}
\note{
This function is inspired by \code{\link[dlnm:plot.crosspred]{dlnm::plot.crosspred()}} (Gasparrini 2011). It has been adapted to work in a Bayesian framework within the \pkg{bdlnm} package.
}
\examples{
# Set exposure-response and lag-response spline parameters
 dlnm_var <- list(
   var_prc = c(10, 75, 90),
   var_fun = "ns",
   lag_fun = "ns",
   max_lag = 21,
   lagnk = 3
 )


# Set cross-basis parameters
 argvar <- list(fun = dlnm_var$var_fun,
                knots = stats::quantile(london$tmean,
                                 dlnm_var$var_prc/100, na.rm = TRUE),
                Bound = range(london$tmean, na.rm = TRUE))

 arglag <- list(fun = dlnm_var$lag_fun,
                knots = dlnm::logknots(dlnm_var$max_lag, nk = dlnm_var$lagnk))

 # Create crossbasis
 cb <- dlnm::crossbasis(london$tmean, lag = dlnm_var$max_lag, argvar, arglag)

 # Seasonality of mortality time series
 seas <- splines::ns(london$date, df = round(8 * length(london$date) / 365.25))

 # Prediction values (equidistant points)
 temp <- round(seq(min(london$tmean), max(london$tmean), by = 0.1), 1)

 # Fit the model
 mod <- bdlnm(mort_75plus ~ cb + factor(dow) + seas, basis = cb, data = london, family = "poisson")

 # Prediction
 cpred <- bcrosspred(mod, cb, at = temp)

 # Perform the plots:

 #Overall
 plot(cpred, "overall", xlab = "Temperature (ºC)", ylab = "Relative Risk", col = 4,
 main="Overall", log = "y")

 #3-D plot
 plot(cpred, "3d", zlab = "Relative risk", col = 4, lphi = 60, cex.axis = 0.6,
 xlab = "Temperature (ºC)", main = "3D graph of temperature effect")

 #Contour
 plot(cpred, "contour", xlab = "Temperature (ºC)", ylab = "Lag", main="Contour plot")

 #Slices (for a high temperature)
 htemp <- 23
 plot(cpred , "slices", var = htemp, col=3, ylab="RR",
 main=paste0("Association for a high temperature (", htemp, "ºC)"))

 #Slices (for lag 0)
 plot(cpred , "slices", lag = 0, col=4, ylab="RR", main=paste0("Association at Lag 0"))


}
\references{
Gasparrini A. Distributed lag linear and non-linear models in R: the package dlnm. Journal of Statistical Software. 2011; 43(8):1-20.
}
\seealso{
\code{\link[=bcrosspred]{bcrosspred()}} to predict exposure–lag–response associations for a \code{"bdlnm"} object,

\code{\link[=bdlnm]{bdlnm()}} to fit a Bayesian distributed lag non-linear model (\code{"bdlnm"}).
}
\author{
Pau Satorra, Marcos Quijal.
}
